--- a/.github/workflows/test.yml
+++ b/.github/workflows/test.yml
@@ -32,7 +32,7 @@ jobs:
           args: --go ${{ matrix.go-version }}
 
       - name: Run tests
-        run: go test -covermode=count -coverprofile=cover.out ./...
+        run: go test -covermode=count -coverprofile=coverage.out ./...
 
       - name: Convert coverage.out to coverage.lcov
         uses: jandelgado/gcov2lcov-action@v1