--- a/Makefile
+++ b/Makefile
@@ -22,14 +22,6 @@ lint:
 	@golangci-lint run ./...
 .PHONY: lint
 
-# Generate Go files
-generate:
-	@echo "==> Generating"
-	@go install -modfile go.tools.mod github.com/a8m/syncmap
-	@go generate
-	@echo "==> Done"
-.PHONY: generate
-
 # Run benchmarks
 bench:
 	@go test -bench . ./...