--- a/go.mod
+++ b/go.mod
@@ -7,6 +7,7 @@ require (
 	github.com/cactus/go-statsd-client/v5 v5.1.0
 	github.com/hamba/logger/v2 v2.8.0
 	github.com/prometheus/client_golang v1.20.5
+	github.com/puzpuzpuz/xsync v1.5.2
 	github.com/stretchr/testify v1.10.0
 	github.com/valyala/fastrand v1.1.0
 )