--- a/registry.go
+++ b/registry.go
@@ -4,6 +4,8 @@ import (
 	"errors"
 	"sync"
 	"time"
+
+	"github.com/puzpuzpuz/xsync"
 )
 
 type registry struct {
@@ -75,14 +77,18 @@ func (r *registry) SubStatter(parent *Statter, prefix string, tags []Tag) *Statt
 	}
 
 	s := &Statter{
-		cfg:    parent.cfg,
-		reg:    r,
-		r:      parent.r,
-		hr:     parent.hr,
-		tr:     parent.tr,
-		pool:   parent.pool,
-		prefix: name,
-		tags:   tags,
+		cfg:        parent.cfg,
+		reg:        r,
+		r:          parent.r,
+		hr:         parent.hr,
+		tr:         parent.tr,
+		pool:       parent.pool,
+		prefix:     name,
+		tags:       tags,
+		counters:   xsync.NewMapOf[*Counter](),
+		gauges:     xsync.NewMapOf[*Gauge](),
+		histograms: xsync.NewMapOf[*Histogram](),
+		timings:    xsync.NewMapOf[*Timing](),
 	}
 	r.statters[k.SafeString()] = s