## Goal of this PR

This switches the generate maps with generics. The performance is largely the same, but now with 100% less generation:

```
name                           old time/op    new time/op    delta
Key-8                            9.04ns ±25%    9.00ns ±28%     ~     (p=0.971 n=10+10)
Statter_Counter-8                89.6ns ±10%    80.1ns ± 1%  -10.59%  (p=0.000 n=10+9)
Statter_Gauge-8                  84.6ns ±11%    76.7ns ± 2%   -9.41%  (p=0.001 n=10+9)
Statter_Histogram-8               138ns ± 2%     142ns ± 1%   +2.91%  (p=0.000 n=9+10)
Statter_Timing-8                  141ns ± 2%     145ns ± 2%   +2.80%  (p=0.000 n=10+10)
Statter_PrometheusHistogram-8     390ns ± 4%     392ns ± 6%     ~     (p=0.912 n=10+10)
Statter_PrometheusTiming-8        380ns ± 5%     367ns ± 8%     ~     (p=0.109 n=10+10)

name                           old alloc/op   new alloc/op   delta
Key-8                             0.00B          0.00B          ~     (all equal)
Statter_Counter-8                 0.00B          0.00B          ~     (all equal)
Statter_Gauge-8                   0.00B          0.00B          ~     (all equal)
Statter_Histogram-8               0.00B          0.00B          ~     (all equal)
Statter_Timing-8                  0.00B          0.00B          ~     (all equal)
Statter_PrometheusHistogram-8     0.00B          0.00B          ~     (all equal)
Statter_PrometheusTiming-8        0.00B          0.00B          ~     (all equal)

name                           old allocs/op  new allocs/op  delta
Key-8                              0.00           0.00          ~     (all equal)
Statter_Counter-8                  0.00           0.00          ~     (all equal)
Statter_Gauge-8                    0.00           0.00          ~     (all equal)
Statter_Histogram-8                0.00           0.00          ~     (all equal)
Statter_Timing-8                   0.00           0.00          ~     (all equal)
Statter_PrometheusHistogram-8      0.00           0.00          ~     (all equal)
Statter_PrometheusTiming-8         0.00           0.00          ~     (all equal)
```

<!--
Fixes #
-->

## How did I test it?

<!-- A brief description the steps taken to test this pull request. -->